test(ts-browser): share `runBenchmark`
browser and native now follow the same logic.
