+++ crypto-ffi/bindings/js/packages/native/benches/utils.ts
@@ -1,10 +1,7 @@
-import {
-    logResults,
-    setup as sharedSetup,
-} from "../../../shared/benches/utils";
+import { setup as sharedSetup } from "../../../shared/benches/utils";
 import { sharedTeardown as tinybenchTeardown } from "../shared/utils";
 
-async function setup() {
+export async function setup() {
     await sharedSetup();
     if (globalThis.tinybench === undefined) {
         globalThis.tinybench = await import("tinybench");
@@ -15,10 +12,4 @@ async function setup() {
     }
 }
 
-export async function runBenchmark(benchmarkSetup: () => Promise<void>) {
-    await setup();
-    await benchmarkSetup();
-    console.log(`Starting ${bench.name}`);
-    await bench.run();
-    await logResults(bench.name, bench.table());
-}
+export async function teardown() {}
